Login i logout dodati, testirane rute

+++ aplikacija/app/Http/Controllers/EventController.php
@@ -17,56 +17,56 @@ class EventController extends Controller
         return EventResource::collection($events);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(Request $request)
-    {
-        $validated = $request->validate([
-            'title' => 'required|string|max:255',
-            'description' => 'required|string',
-            'date' => 'required|date',
-            'user_id' => 'required|exists:users,id',
-        ]);
-
-        $event = Event::create($validated);
-
-        return new EventResource($event);
-    }
-
-    /**
-     * Display the specified resource.
-     */
     public function show(Event $event)
     {
         $event->load('user'); // Učitajte povezanog korisnika
         return new EventResource($event);
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, Event $event)
+    public function store(Request $request)
     {
         $validated = $request->validate([
-            'title' => 'sometimes|string|max:255',
-            'description' => 'sometimes|string',
-            'date' => 'sometimes|date',
-            'user_id' => 'sometimes|exists:users,id',
+            'title' => 'required|string|max:255',
+            'description' => 'nullable|string',
+            'start_time' => 'required|date',
+            'end_time' => 'required|date|after:start_time',
+            'user_id' => 'required|exists:users,id',
+            'category_id' => 'required|exists:event_categories,id',
         ]);
 
-        $event->update($validated);
+        $event = Event::create($validated);
 
-        return new EventResource($event);
+        return $event;
+        
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Event $event)
-    {
-        $event->delete();
+   
 
-        return response()->json(null, 204);
+    public function destroy($id)
+    {
+        $event = Event::find($id);
+        if (!$event) {
+            return response()->json(['error' => 'Event not found'], 404);
+        }
+        $event->delete();
+        return response()->json($event, 200);
+    }
+
+    public function update(Request $request, Event $event)
+    {
+        $validated = $request->validate([
+            'title' => 'sometimes|required|string|max:255',
+            'description' => 'sometimes|required|nullable|string',
+            'start_time' => 'sometimes|required|date',
+            'end_time' => 'sometimes|required|date|after:start_time',
+            'user_id' => 'sometimes|required|exists:users,id',
+            'category_id' => 'sometimes|required|exists:event_categories,id',
+        ]);
+
+        $event->update($validated);
+        return $event;
     }
 }
